refactor: improving use case destroy hero

+++ app/use_cases/destroy_hero/steps/destroy.rb
@@ -4,6 +4,9 @@ class DestroyHero::Steps::Destroy
   include Dry::Monads[:result]
 
   def call(hero)
+    raise StandardError.new(I18n.t(:hero_working, hero: hero.name)) if hero.working?
+
+    hero.remove_from_index!
     hero.destroy!
   end
 end
